pyproject.toml
[tool.poetry]
name = "romanenko-uchit-bot"
version = "0.1.0"
description = ""
authors = ["romanenko_vova <bonyshd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = {extras = ["job-queue"], version = "^21.4"}
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
